Successfully managed to make calculator file -HarshitRSethi
